Cross-build for Scala.js
