+++ ast/js/src/main/scala/jawn/ast/JParserPlatform.scala
@@ -1,0 +1,4 @@
+package org.typelevel.jawn
+package ast
+
+private[jawn] trait JParserPlatform
