+++ ast/js/src/test/scala/jawn/ast/ParseCheckPlatform.scala
@@ -1,0 +1,4 @@
+package org.typelevel.jawn
+package ast
+
+private[jawn] trait AstCheckPlatform
