+++ parser/js/src/main/scala/jawn/SyntaxPlatform.scala
@@ -1,0 +1,3 @@
+package org.typelevel.jawn
+
+private[jawn] trait SyntaxPlatform
