+++ parser/js/src/test/scala/jawn/JNumIndexCheckPlatform.scala
@@ -1,0 +1,4 @@
+package org.typelevel.jawn
+package parser
+
+private[jawn] trait JNumIndexCheckPlatform
