+++ util/js/src/test/scala/jawn/util/SliceCheckPlatform.scala
@@ -1,0 +1,4 @@
+package org.typelevel.jawn
+package util
+
+private[jawn] trait SliceCheckPlatform
